Fix compiler warnings

+++ src/Battlescape/SkillMenuState.cpp
@@ -55,7 +55,7 @@ SkillMenuState::SkillMenuState(BattleAction *action, int x, int y) : ActionMenuS
 	// Set palette
 	_game->getSavedGame()->getSavedBattle()->setPaletteByDepth(this);
 
-	for (int i = 0; i < std::size(_actionMenu); ++i)
+	for (int i = 0; i < (int)std::size(_actionMenu); ++i)
 	{
 		_actionMenu[i] = new ActionMenuItem(i, _game, x, y);
 		add(_actionMenu[i]);
